Исправлены тесты для асинхронных операций
