+++ src/store/user/user.test.js
@@ -1,26 +1,11 @@
-import axios from 'axios';
-import MockAdapter from 'axios-mock-adapter';
-import {DataType, LoadingStatus} from "../../const";
-import {createAPI} from '../../services/api';
-import {adaptDataToClient} from "../../utils/adaptDataToClient";
 import {changeAuthorizationStatus, setUser} from "../action-creator";
-import {ActionType} from "../actions";
-import {checkAuthorization, login, logout} from "../api-actions";
 import {user} from "./user";
-import * as redux from 'react-redux';
 
-const api = createAPI(() => {});
 const fakeUser = {
   id: `1`,
   name: `User`,
   image: ``,
   email: ``
-};
-const fakeUserFromServer = {
-  "id": 1,
-  "name": `User`,
-  "avatar_url": ``,
-  "email": ``
 };
 
 describe(`User reducers work correctly`, () => {
@@ -69,149 +54,4 @@ describe(`User reducers work correctly`, () => {
     expect(user(initialState, setUser(fakeUser))).toEqual(expectedState);
   });
 
-});
-describe(`Async operations work correctly: user`, () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-  it(`Should make a correct API call to /login`, () => {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const adaptData = jest.spyOn(adaptDataToClient, DataType.USER);
-    const fakeFormData = {email: `asd@test.ru`, password: `testpass`};
-    const loginHandler = login(fakeFormData);
-
-    loginHandler(dispatch, () => {}, api);
-    expect(dispatch).toHaveBeenCalledTimes(1);
-    expect(dispatch).toHaveBeenNthCalledWith(1, {
-      type: ActionType.CHANGE_SENDING_DATA_STATUS,
-      payload: LoadingStatus.FETCHING,
-    });
-
-    apiMock
-      .onPost(`/login`)
-      .reply(200, fakeUserFromServer);
-
-    axios
-      .get(`/login`)
-      .then(() => {
-        expect(adaptData).toHaveBeenCalledTimes(1);
-        expect(adaptData).toHaveBeenCalledWith(fakeUserFromServer);
-
-        expect(dispatch).toHaveBeenCalledTimes(4);
-        expect(dispatch).toHaveBeenNthCalledWith(1, {
-          type: ActionType.SET_USER,
-          payload: fakeUser,
-        });
-        expect(dispatch).toHaveBeenNthCalledWith(2, {
-          type: ActionType.CHANGE_SENDING_DATA_STATUS,
-          payload: LoadingStatus.SUCCESS,
-        });
-        expect(dispatch).toHaveBeenNthCalledWith(3, {
-          type: ActionType.CHANGE_AUTHORIZATION_STATUS,
-          payload: true,
-        });
-        expect(dispatch).toHaveBeenNthCalledWith(4, {
-          type: ActionType.REDIRECT_TO_ROUTE,
-          payload: `/`,
-        });
-      })
-      .catch(() => {
-        expect(dispatch).toHaveBeenCalledTimes(2);
-        expect(dispatch).toHaveBeenNthCalledWith(1, {
-          type: ActionType.CHANGE_SENDING_DATA_STATUS,
-          payload: LoadingStatus.FAILURE,
-        });
-        expect(dispatch).toHaveBeenNthCalledWith(2, {
-          type: ActionType.CHANGE_AUTHORIZATION_STATUS,
-          payload: false,
-        });
-      });
-  });
-  it(`Should make a correct API call to /login for check authorization`, () => {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const adaptData = jest.spyOn(adaptDataToClient, DataType.USER);
-    const checkAuthorizationHandler = checkAuthorization();
-
-    apiMock
-      .onGet(`/login`)
-      .reply(200, fakeUserFromServer);
-
-    checkAuthorizationHandler(dispatch, () => {}, api)
-      .then(() => {
-        expect(adaptData).toHaveBeenCalledTimes(1);
-        expect(adaptData).toHaveBeenCalledWith(fakeUserFromServer);
-
-        expect(dispatch).toHaveBeenCalledTimes(2);
-        expect(dispatch).toHaveBeenNthCalledWith(1, {
-          type: ActionType.SET_USER,
-          payload: fakeUser,
-        });
-        expect(dispatch).toHaveBeenNthCalledWith(2, {
-          type: ActionType.CHANGE_AUTHORIZATION_STATUS,
-          payload: true,
-        });
-      })
-      .catch(() => {
-        expect(dispatch).toHaveBeenCalledTimes(1);
-        expect(dispatch).toHaveBeenNthCalledWith(1, {
-          type: ActionType.CHANGE_AUTHORIZATION_STATUS,
-          payload: false,
-        });
-      });
-  });
-  it(`Should make a correct API call to /logout`, () => {
-    const apiMock = new MockAdapter(api);
-    const dispatch = jest.fn();
-    const logoutHandler = logout();
-
-    logoutHandler(dispatch, () => {}, api);
-    expect(dispatch).toHaveBeenCalledTimes(1);
-    expect(dispatch).toHaveBeenNthCalledWith(1, {
-      type: ActionType.CHANGE_SENDING_DATA_STATUS,
-      payload: LoadingStatus.FETCHING,
-    });
-
-    apiMock
-      .onPost(`/logout`)
-      .reply(200);
-
-    axios
-      .get(`/logout`)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(5);
-        expect(dispatch).toHaveBeenNthCalledWith(1, {
-          type: ActionType.SET_USER,
-          payload: null,
-        });
-        expect(dispatch).toHaveBeenNthCalledWith(2, {
-          type: ActionType.CHANGE_SENDING_DATA_STATUS,
-          payload: LoadingStatus.SUCCESS,
-        });
-        expect(dispatch).toHaveBeenNthCalledWith(3, {
-          type: ActionType.CHANGE_AUTHORIZATION_STATUS,
-          payload: false,
-        });
-        expect(dispatch).toHaveBeenNthCalledWith(4, {
-          type: ActionType.CHANGE_POSTER_LOADING_STATUS,
-          payload: LoadingStatus.INITIAL,
-        });
-        expect(dispatch).toHaveBeenNthCalledWith(5, {
-          type: ActionType.CHANGE_FILM_LOADING_STATUS,
-          payload: LoadingStatus.INITIAL,
-        });
-      })
-      .catch(() => {
-        expect(dispatch).toHaveBeenCalledTimes(2);
-        expect(dispatch).toHaveBeenNthCalledWith(1, {
-          type: ActionType.CHANGE_SENDING_DATA_STATUS,
-          payload: LoadingStatus.FAILURE,
-        });
-        expect(dispatch).toHaveBeenNthCalledWith(2, {
-          type: ActionType.CHANGE_AUTHORIZATION_STATUS,
-          payload: true,
-        });
-      });
-  });
 });
